Refactor save/update playlists, modularize helpers, add time
